Move items that do not fit into single aggregate into Common
Gender, SetType, and MatchFormat
