+++ src/TournamentManagement/TournamentManagement.Domain.UnitTests/MatchAggregate/MatchScoreTests.cs
@@ -1,8 +1,8 @@
 ﻿using FluentAssertions;
 using System;
 using System.Collections.Generic;
+using TournamentManagement.Domain.Common;
 using TournamentManagement.Domain.MatchAggregate;
-using TournamentManagement.Domain.TournamentAggregate;
 using Xunit;
 
 namespace TournamentManagement.Domain.UnitTests.MatchAggregate
