+++ src/TournamentManagement/TournamentManagement.Domain.UnitTests/PlayerAggregate/PlayerTests.cs
@@ -1,5 +1,6 @@
 ﻿using FluentAssertions;
 using System;
+using TournamentManagement.Domain.Common;
 using TournamentManagement.Domain.PlayerAggregate;
 using Xunit;
