+++ src/TournamentManagement/TournamentManagement.Domain/Common/Gender.cs
@@ -1,0 +1,8 @@
+﻿namespace TournamentManagement.Domain.Common
+{
+	public enum Gender
+	{
+		Male,
+		Female
+	}
+}
