--- src/TournamentManagement/TournamentManagement.Domain/Gender.cs
@@ -1,8 +1,0 @@
-﻿namespace TournamentManagement.Domain
-{
-	public enum Gender
-	{
-		Male,
-		Female
-	}
-}
