+++ src/TournamentManagement/TournamentManagement.Domain/MatchAggregate/MatchScore.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using TournamentManagement.Domain.TournamentAggregate;
+using TournamentManagement.Domain.Common;
 
 namespace TournamentManagement.Domain.MatchAggregate
 {
